chore: remove extra keywords 'struct' and 'union', add casts to eliminate warnings

+++ source/others/FAST_INT64/f128_to_i32_r_minMag.cpp
@@ -87,8 +87,7 @@ f128_to_i32_r_minMag(float128_t a,
     }
 
     sig64 |= UINT64_C(0x0001000000000000);
-    /** @todo Warning   C4244   '=': conversion from 'uint64_t' to 'int32_t', possible loss of data */
-    absZ = sig64 >> shiftDist;
+    absZ = static_cast<int32_t>(sig64 >> shiftDist);
 
     if (
         exact && ((uint64_t)(uint32_t)absZ << shiftDist != sig64)
